+++ src/main/kotlin/net/postchain/rellide/jetbrains/util/FileUtil.kt
@@ -2,6 +2,4 @@ package net.postchain.rellide.jetbrains.util
 
 import com.intellij.openapi.vfs.VirtualFile
 
-fun VirtualFile.normalizedUri(): String {
-    return this.url.replace("file:///", "file:/")
-}
+fun VirtualFile.normalizedUri(): String = this.url.replace("file:///", "file:/")